Added ProductListItem and update menuScreen
