+++ src/components/ProductListItem.tsx
@@ -1,0 +1,43 @@
+import { StyleSheet, Image } from "react-native";
+import { Text, View } from "@/src/components/Themed";
+import Colors from "@/src/constants/Colors";
+import { Product } from "@/types";
+
+type ProductListItemProps = {
+  product: Product;
+};
+
+export default function ProductListItem({ product }: ProductListItemProps) {
+  //todo: add image placeholder if no image is found
+  if (!product.image) product.image = "https://notjustdev-dummy.s3.us-east-2.amazonaws.com/food/default.png";
+  return (
+    <View style={styles.container}>
+      <Image style={styles.image} source={{ uri: product.image }} />
+      <Text style={styles.title}>{product.name}</Text>
+      <Text style={styles.price}>${product.price}</Text>
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: {
+    backgroundColor: "white",
+    padding: 10,
+    borderRadius: 20
+  },
+
+  image: {
+    width: "100%",
+    aspectRatio: 1
+  },
+
+  title: {
+    fontSize: 18,
+    fontWeight: "600",
+    marginVertical: 10
+  },
+  price: {
+    color: Colors.light.tint,
+    fontWeight: "bold"
+  }
+});
